src/requirements.txt
pandas==1.5.3
streamlit==1.31.1
SQLAlchemy
pandasai==2.0
langchain
pygwalker
psycopg2-binary
pyjwt
python-gettext
polars